refactor: split _feature_chaining into ModelReader and FeatureChainingResolver
